+++ NEWS.md
@@ -22,6 +22,7 @@
 
 ## Bug fixes
 
+- Fixed bug in `weighted_posteriors()` to properly weight Intercept-only `BFBayesFactor` models.
 - Fixed bug in `weighted_posteriors()` when models have very low posterior probability ( #286 ).
 - Fixed bug in `describe_posterior()`, `rope()` and `equivalence_test()` for *brmsfit* models with monotonic effect.
 - Fixed issues related to latest changes in `as.data.frame.brmsfit()` from the *brms* package.